refactoring: isloating drain mode code
