+++ pulsar/reader_drain.go
@@ -1,0 +1,46 @@
+// Licensed to the Apache Software Foundation (ASF) under one
+// or more contributor license agreements.  See the NOTICE file
+// distributed with this work for additional information
+// regarding copyright ownership.  The ASF licenses this file
+// to you under the Apache License, Version 2.0 (the
+// "License"); you may not use this file except in compliance
+// with the License.  You may obtain a copy of the License at
+//
+//   http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing,
+// software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+// KIND, either express or implied.  See the License for the
+// specific language governing permissions and limitations
+// under the License.
+
+package pulsar
+
+// ReaderWithDrainMode is a Reader that additionally supports drain mode.
+type ReaderWithDrainMode interface {
+	Reader
+
+	// EnterDrainMode stops sending permit flow requests to the broker, preventing
+	// delivery of any new messages while allowing the reader to process existing
+	// messages in the client buffer.
+	EnterDrainMode() error
+
+	// ExitDrainMode resumes normal message flow by allowing the reader to send
+	// permit flow requests to the broker for new message delivery.
+	ExitDrainMode() error
+}
+
+func (r *reader) EnterDrainMode() error {
+	r.Lock()
+	defer r.Unlock()
+
+	return r.c.EnterDrainMode()
+}
+
+func (r *reader) ExitDrainMode() error {
+	r.Lock()
+	defer r.Unlock()
+
+	return r.c.ExitDrainMode()
+}
