+++ pulsar/reader_impl.go
@@ -223,20 +223,6 @@ func (r *reader) SeekByTime(time time.Time) error {
 	return r.c.SeekByTime(time)
 }
 
-func (r *reader) EnterDrainMode() error {
-	r.Lock()
-	defer r.Unlock()
-
-	return r.c.EnterDrainMode()
-}
-
-func (r *reader) ExitDrainMode() error {
-	r.Lock()
-	defer r.Unlock()
-
-	return r.c.ExitDrainMode()
-}
-
 func (r *reader) GetLastMessageID() (MessageID, error) {
 	consumers := r.c.partitionConsumers()
 	if len(consumers) > 1 {
